+++ app/src/main/java/ua/com/radiokot/photoprism/features/gallery/data/storage/SearchPreferences.kt
@@ -5,4 +5,5 @@ import io.reactivex.rxjava3.subjects.BehaviorSubject
 interface SearchPreferences {
     val showPeople: BehaviorSubject<Boolean>
     val showAlbums: BehaviorSubject<Boolean>
+    val showAlbumFolders: BehaviorSubject<Boolean>
 }